fft example update

+++ src/libs/maxiFFT.cpp
@@ -112,7 +112,7 @@ vector<float> & maxiFFT::magsToDB() {
 
 float maxiFFT::spectralFlatness() {
 	float geometricMean=0, arithmaticMean=0;
-	for(int i=0; i < bins; i++) {
+	for(size_t i=0; i < bins; i++) {
 		if (magnitudes[i] != 0)
 			geometricMean += logf(magnitudes[i]);
 		arithmaticMean += magnitudes[i];
@@ -124,7 +124,7 @@ float maxiFFT::spectralFlatness() {
 
 float maxiFFT::spectralCentroid() {
 	float x=0, y=0;
-	for(int i=0; i < bins; i++) {
+	for(size_t i=0; i < bins; i++) {
 		x += fabs(magnitudes[i]) * i;
 		y += fabs(magnitudes[i]);
 	}
